Revert "minor refactoring"
This reverts commit 8ddcedb0f61e4db6e27afc6a294775cb8490b3d4.
