+++ packages/events/src/service/users/api.ts
@@ -15,6 +15,9 @@ import {
   FullDocumentPath,
   UUID,
   IUserName,
+  UserOrSystem,
+  TokenUserType,
+  logger,
   SystemRole
 } from '@opencrvs/commons'
 import { env } from '@events/environment'
@@ -41,7 +44,7 @@ type UserAPIResult = {
 export async function getUser(
   userId: string,
   token: string
-): Promise<UserAPIResult | undefined> {
+): Promise<UserAPIResult> {
   const res = await fetch(joinUrl(env.USER_MANAGEMENT_URL, 'getUser').href, {
     method: 'POST',
     body: JSON.stringify({ userId }),
@@ -52,7 +55,9 @@ export async function getUser(
   })
 
   if (!res.ok) {
-    return undefined
+    throw new Error(
+      `Unable to retrieve user details. Error: ${res.status} status received`
+    )
   }
 
   return res.json() as Promise<UserAPIResult>
@@ -72,7 +77,7 @@ type SystemAPIResult = {
 export async function getSystem(
   systemId: string,
   token: string
-): Promise<SystemAPIResult | undefined> {
+): Promise<SystemAPIResult> {
   const res = await fetch(joinUrl(env.USER_MANAGEMENT_URL, 'getSystem').href, {
     method: 'POST',
     body: JSON.stringify({ systemId }),
@@ -83,8 +88,53 @@ export async function getSystem(
   })
 
   if (!res.ok) {
-    return undefined
+    throw new Error(
+      `Unable to retrieve system details. Error: ${res.status} status received`
+    )
   }
 
   return res.json() as Promise<SystemAPIResult>
+}
+
+export async function getUserOrSystem(
+  id: string,
+  token: string
+): Promise<UserOrSystem | undefined> {
+  try {
+    const user = await getUser(id, token)
+
+    return {
+      type: TokenUserType.enum.user,
+      id: user.id,
+      name: user.name,
+      role: user.role,
+      signature: user.signature ? user.signature : undefined,
+      avatar: user.avatar?.data ? user.avatar.data : undefined,
+      primaryOfficeId: user.primaryOfficeId,
+      fullHonorificName: user.fullHonorificName
+        ? user.fullHonorificName
+        : undefined
+    }
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  } catch (_) {
+    logger.info(`No user found for id: ${id}. Will look for a system instead.`)
+  }
+
+  try {
+    const system = await getSystem(id, token)
+
+    return {
+      type: TokenUserType.enum.system,
+      id,
+      name: system.name,
+      role: system.type
+    }
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  } catch (e) {
+    logger.info(
+      `No system found for id: ${id}. User/system has probably been removed. Will return undefined.`
+    )
+  }
+
+  return
 }
